renamed duke to g
