+++ src/main/java/g/commands/AddCommand.java
@@ -1,9 +1,9 @@
-package duke.commands;
+package g.commands;
 
-import duke.storage.Storage;
-import duke.tasks.Task;
-import duke.tasks.TaskList;
-import duke.ui.Ui;
+import g.storage.Storage;
+import g.tasks.Task;
+import g.tasks.TaskList;
+import g.ui.Ui;
 
 public class AddCommand extends Command {
     private final Task task;
