+++ src/main/java/g/commands/ListCommand.java
@@ -1,8 +1,8 @@
-package duke.commands;
+package g.commands;
 
-import duke.storage.Storage;
-import duke.tasks.TaskList;
-import duke.ui.Ui;
+import g.storage.Storage;
+import g.tasks.TaskList;
+import g.ui.Ui;
 
 public class ListCommand extends Command {
