+++ src/main/java/g/tasks/Deadline.java
@@ -1,4 +1,4 @@
-package duke.tasks;
+package g.tasks;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
