+++ src/main/java/g/tasks/Task.java
@@ -1,4 +1,4 @@
-package duke.tasks;
+package g.tasks;
 public abstract class Task {
     protected String description;
     protected boolean isDone;
