+++ src/main/java/g/tasks/TaskList.java
@@ -1,4 +1,4 @@
-package duke.tasks;
+package g.tasks;
 
 import java.util.ArrayList;
